requirements.txt
absl-py==2.1.0
accelerate==0.20.3
ace_tools==0.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
autograd==1.7.0
autoray==0.7.0
av==12.3.0
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==1.0.1
branca==0.7.2
cachelib==0.13.0
cachetools==5.5.0
captcha==0.6.0
catalogue==2.0.10
catboost==1.2.5
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
confection==0.1.5
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cryptography==43.0.1
cvlib==0.2.7
cvzone==1.6.1
cycler==0.12.1
cymem==2.0.8
dask==2024.7.0
dataclasses-json==0.6.7
deap==1.4.1
debugpy==1.8.1
decorator==4.4.2
deepface==0.0.93
diffusers==0.17.1
distributed==2024.7.0
Django==5.0.6
dnspython==2.7.0
docstring_parser==0.16
docx2txt==0.8
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.1
executing==2.0.1
extra-streamlit-components==0.1.71
facenet-pytorch==2.6.0
Faker==27.0.0
fastai==2.7.18
fastcore==1.7.19
fastdownload==0.0.7
fastjsonschema==2.21.1
fastprogress==1.0.3
fer==22.5.1
ffmpeg==1.4
filelock==3.14.0
fire==0.7.0
fitz==0.0.1.dev2
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-Sessionstore==0.4.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
flexcache==0.3
flexparser==0.3.1
folium==0.17.0
fonttools==4.52.4
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.3.1
gast==0.6.0
gdown==5.2.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
graphviz==0.20.3
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
gTTS==2.5.3
gunicorn==23.0.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.11.1
httplib2==0.22.0
huggingface-hub==0.15.1
hyperframe==5.2.0
HyperPyYAML==1.2.2
idna==3.10
ifaddr==0.2.0
imageio==2.36.0
imageio-ffmpeg==0.5.1
imbalanced-learn==0.12.3
importlib-metadata==6.11.0
imutils==0.5.4
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.25.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.15.0
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langcodes==3.4.1
langdetect==1.0.9
langsmith==0.1.135
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
libretranslatepy==2.1.1
librosa==0.10.2.post1
lightgbm==4.5.0
llvmlite==0.43.0
locket==1.0.0
looseversion==1.3.0
lxml==5.2.2
lz4==4.3.3
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mathparse==0.1.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
missingno==0.5.2
mkl==2021.4.0
ml_dtypes==0.5.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
mtcnn==1.0.0
multidict==6.1.0
murmurhash==1.0.10
mutagen==1.47.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==9.1.0
namex==0.0.8
narwhals==1.5.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.7
packaging==23.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
patchify==0.2.3
pathlib==1.0.1
patsy==0.5.6
pdfminer==20191125
pdfminer.six==20250327
PennyLane==0.39.0
PennyLane_Lightning==0.39.0
Pillow==9.5.0
Pint==0.24.3
platformdirs==4.2.2
playsound==1.3.0
plotly==5.22.0
pooch==1.8.2
preshed==3.0.9
proglog==0.1.10
progressbar==2.5
prompt_toolkit==3.0.45
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.3
prov==2.0.1
psutil==6.1.1
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pyAudioAnalysis==0.3.14
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.3
pydantic-settings==2.5.2
pydantic_core==2.27.1
pydeck==0.9.1
pydot==2.0.0
pydub==0.25.1
pyee==12.1.1
pyenchant==3.2.2
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.10.1
pylibsrtp==0.10.0
pymongo==3.13.0
Pympler==1.1
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.0
PyPDF2==3.0.1
pypiwin32==223
pyresparser==1.0.6
pyrsistent==0.20.0
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.0.3
rdflib==6.3.2
referencing==0.35.1
regex==2024.5.15
replicate==1.0.4
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
rustworkx==0.15.1
safetensors==0.4.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
shiboken6==6.8.0.2
simplejson==3.19.2
six==1.16.0
skrebate==0.62
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.2
SpeechRecognition==3.11.0
spektral==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.0
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.2
stopit==1.1.2
streamlit==1.44.0
streamlit-audiorecorder==0.0.6
streamlit-authenticator==0.4.1
streamlit-option-menu==0.4.0
streamlit-webrtc==0.47.9
streamlit_mic_recorder==0.0.8
sympy==1.13.1
tabulate==0.9.0
tbb==2021.13.1
tblib==3.0.0
tenacity==8.3.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.15.0
tensorflow-federated==0.1.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
textblob==0.18.0.post0
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.1
torch==1.12.1
tornado==6.4
TPOT==0.12.2
tqdm==4.67.1
traitlets==5.14.3
traits==6.3.2
transformers==4.30.2
translate==3.6.1
twilio==9.4.3
typeguard==2.13.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.13.0
tzdata==2024.1
tzlocal==4.3.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
validators==0.33.0
vit-keras==0.1.2
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wordcloud==1.9.3
wrapt==1.14.1
WTForms==3.1.2
xgboost==2.0.3
xlsx2html==0.6.1
xyzservices==2024.6.0
yarl==1.15.3
zict==3.0.0
zipp==3.19.2
